move app folder to addon folder and add reexports
