+++ addon/helpers/task.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+
+function taskHelper([task, ...args]) {
+  return task._curry(...args);
+}
+
+export default Ember.Helper.helper(taskHelper);
+
