+++ addon/initializers/ember-concurrency.js
@@ -1,0 +1,8 @@
+// This initializer exists only to make sure that the following
+// imports happen before the app boots.
+import 'ember-concurrency';
+
+export default {
+  name: 'ember-concurrency',
+  initialize: function() {}
+};
