+++ app/helpers/perform.js
@@ -1,9 +1,1 @@
-import Ember from 'ember';
-import { taskHelperClosure } from 'ember-concurrency/-helpers';
-
-export function performHelper(args, hash) {
-  return taskHelperClosure('perform', 'perform', args, hash);
-}
-
-export default Ember.Helper.helper(performHelper);
-
+export { default, perform } from 'ember-concurrency/helpers/perform';
