+++ app/helpers/task.js
@@ -1,8 +1,1 @@
-import Ember from 'ember';
-
-function taskHelper([task, ...args]) {
-  return task._curry(...args);
-}
-
-export default Ember.Helper.helper(taskHelper);
-
+export { default, task } from 'ember-concurrency/helpers/task';
